various changes to `cpu` to make it work on my laptop
for some reason, my laptop doesn't populate the `CPU MHz` field of
`/proc/cpuinfo`, so `fmt-cpu-freq` doesn't work. simple fix: don't call
`fmt-cpu-freq` in `cpu-modeline`.

this module was super inconsistent about which format functions took a modeline
as an ignored argument. i have standardized that by making all of the `fmt-*`
methods take an optional ignored argument for the modeline. this is because
stumpwm invokes modeline-formatters with the modeline as their first argument,
but it is convenient to be able to call them with no arguments.